+++ packages/media-avatar-picker/stories/image-cropper/image-cropper-story.tsx
@@ -14,7 +14,7 @@ storiesOf('Stateless image cropper', {})
     <ImageCropper imageSource={tallImage} scale={0.14} top={-50} left={-115} />
   ))
   .add('with custom container size', () => (
-    <ImageCropper imageSource={tallImage} imageWidth={naturalWidth} scale={0.14} top={-50} left={-115} containerSize={400} />
+    <ImageCropper imageSource={tallImage} imageWidth={naturalWidth} scale={0.14} top={-50} left={-115} containerDimensions={{width: 400, height: 400}} />
   ))
   .add('with circular mask', () => (
     <ImageCropper imageSource={tallImage} imageWidth={naturalWidth} scale={0.08} top={-70} left={-90} isCircularMask={true} />