+++ packages/media-avatar-picker/stories/image-navigator/styled.ts
@@ -1,0 +1,13 @@
+/* tslint:disable:variable-name */
+import styled from 'styled-components';
+
+export const CoverWrapper = styled.div`
+  background: #ccc;
+  width: 800px;
+`;
+
+export const Image = styled.img`
+  position: absolute;
+  top: 0;
+  left: 300px;
+`;